+++ CHANGELOG.md
@@ -8,3 +8,10 @@ Fix Some Function
 Add Yur Loading
 ## 0.0.5
 Fix Fonts
+## 0.0.6
+Fixing context
+## 0.0.7
+Fixing Context
+## 0.0.9
+Add PrimaryBlue
+